requirements.txt
flask
flask-sqlalchemy
flask-restx
tqdm
